pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "velixar"
version = "0.1.0"
description = "Persistent memory infrastructure for AI applications"
readme = "README.md"
license = "MIT"
authors = [{ name = "Velixar", email = "sdk@velixarai.com" }]
keywords = ["ai", "memory", "llm", "langchain", "vector", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
langchain = ["langchain-core>=0.1.0"]
llamaindex = ["llama-index-core>=0.10.0"]
all = ["velixar[langchain,llamaindex]"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-httpx>=0.22.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://velixarai.com"
Documentation = "https://docs.velixarai.com"
Repository = "https://github.com/velixar/velixar-python"

[tool.hatch.build.targets.wheel]
packages = ["velixar"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
